Camera scrolling and resizing with screen
Imported Auto Letterbox plugin to ensure constant aspect ratio of game. Added camera controller to pan the camera to follow the player, stopping panning at the edges of scenes. Platformer canvas is now handled by the CameraController instead of the GameManager because the canvas is now in world space and follows the camera as a child object (this ensures that UI elements are not rendered within Auto Letterbox zones)